third_party/blink/web_tests/platform/generic/fast/dom/Document/CaretRangeFromPoint/caretRangeFromPoint-with-first-letter-style-expected.txt

PASS range.startOffset is 0
PASS range.startOffset is 1
PASS range.startOffset is 2
PASS range.startOffset is 3
PASS range.startOffset is 4
PASS range.startOffset is 5
PASS range.startOffset is 6
PASS successfullyParsed is true

TEST COMPLETE
